% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_fdr}
\alias{plot_fdr}
\alias{plot_fnr}
\title{plot_fnr and plot_fdr}
\usage{
plot_fdr(x)

plot_fnr(x)
}
\arguments{
\item{x}{object returned by power.tibs()}
}
\value{
plot of fnr or fdr against sample size (number of pairs) is drawn on current graphics device
}
\description{
Plot FDR or FNR results of power.tibs()
}
\author{
Chris Wallace
}

