% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{tibs.plot}
\alias{tibs.plot}
\title{tibs.plot}
\usage{
tibs.plot(x, y, low = NULL, high = NULL, ...)
}
\arguments{
\item{x}{vector giving position of points on x axis}

\item{y}{matrix, each column to be plotted against x}

\item{low}{optional matrix of lower interval about y}

\item{high}{optional matrix of upper interval about y}

\item{...}{other params passed to plot()}
}
\value{
plot is drawn on current graphics device
}
\description{
Plot FDR or FNR results of power.tibs()
}
\author{
Chris Wallace
}

