% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{tibs.power}
\alias{tibs.power}
\title{tibs.power}
\usage{
tibs.power(norm, ngenes.diff, fold.diff, nsim, sample.sizes, p.thr)
}
\arguments{
\item{norm}{matrix of normalized gene expression values from a reference expt}

\item{ngenes.diff}{number of differentially expressed genes}

\item{fold.diff}{fold difference at differentially expressed genes}

\item{nsim}{number of simulations}

\item{sample.sizes}{number of pairs of samples in simulated experiment, can be a vector}

\item{p.thr}{p value threshold for calling significance, can be a vector}
}
\value{
list summarising the median, 5th and 95th centiles of the false discovery rate (FDR) and false negative rate (FNR)
}
\description{
Estimate power for microarray studies by resampling existing dataset
}
\details{
Uses Tibshirani method (2006)
}
\author{
Chris Wallace
}
\references{
Tibshirani, R. (2006). A simple method for assessing sample sizes in microarray experiments. BMC Bioinformatics 7, 106.
}

